DICOMStructure class split up
