+++ test/test_pyrad_to_sr.py
@@ -5,7 +5,6 @@ import os
 from pydicom import dcmread
 from pydicom.errors import InvalidDicomError
 from pathlib import Path
-from src.Model import DICOMStructuredReport
 from src.Model import ImageLoading
 from src.Model.PatientDictContainer import PatientDictContainer
